requirements.txt
python-dotenv
langchain
langchain-community
langsmith
transformers
torch
pillow
boto3